Adressing PR comments changing visibility of functions and moving/introducing service functions.
