+++ tests/e2e/tests/protoMsgToJSON.go
@@ -1,0 +1,28 @@
+// Copyright (C) 2022-2025, Chain4Travel AG. All rights reserved.
+// See the file LICENSE for licensing terms.
+
+package tests
+
+import (
+	"google.golang.org/protobuf/encoding/protojson"
+	"google.golang.org/protobuf/proto"
+)
+
+/*
+Service function to convert the responses into pretty-printed JSON.
+Only used for debugging and test creation.
+*/
+func protoMessageToJSON(tt *Test, message proto.Message) string {
+	// Pretty-print using protojson.MarshalOptions
+	marshaler := protojson.MarshalOptions{
+		Multiline: true,
+		Indent:    "  ",
+	}
+	jsonData, err := marshaler.Marshal(message)
+	if err != nil {
+		tt.logger.Errorf("Error marshalling: %v", err)
+		return ""
+	}
+	// return the json data as string
+	return string(jsonData)
+}
